[Refactor] rearrange Serice method
[Refactor] 불필요한 코드 삭제 및 product 서비스의 의존적 도메인 책임 분리
