+++ BOOT/product/src/main/java/com/coopang/product/infrastructure/repository/productStock/ProductStockJpaRepository.java
@@ -6,6 +6,6 @@ import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductStockJpaRepository extends JpaRepository<ProductStockEntity, UUID>,
-    ProductStockRepository {
+    ProductStockRepository,ProductStockRepositoryCustom {
     
 }
